% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{key_Program}
\alias{key_Program}
\title{Key Program Identification}
\usage{
key_Program(data, Clustering = 10)
}
\arguments{
\item{data}{A numeric matrix of molecular expression profiles.}

\item{Clustering}{An integer specifying the number of clusters to use.}
}
\value{
A data frame with clustering results.
}
\description{
This function identifies key molecular programs using NMF.
}
\examples{
data <- matrix(rnorm(100), 10, 10)
key_programs <- key_Program(data, Clustering = 5)
}
