% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{normalize}
\alias{normalize}
\title{This function is used to for the normalized matrix}
\usage{
normalize(x)
}
\arguments{
\item{X}{}
}
\value{
normalize_matrix
}
\description{
This function is used to for the normalized matrix
}
