% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{Orthogonal}
\alias{Orthogonal}
\title{Orthogonal Transformation}
\usage{
Orthogonal(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A matrix that has been orthogonally transformed.
}
\description{
This function applies an orthogonal transformation to the given matrix to remove dependency.
}
\examples{
A <- matrix(rnorm(25), 5, 5)
ortho_A <- Orthogonal(A)
}
