% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{PCA_adjust}
\alias{PCA_adjust}
\title{The matrix was debatched using PCA}
\usage{
PCA_adjust(matrix)
}
\arguments{
\item{matrix}{}
}
\value{
Debatching the matrix
}
\description{
The matrix was debatched using PCA
}
