% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{penalty_function_L1}
\alias{penalty_function_L1}
\title{The inner function is used for L1 regularization with cosine similarity algorithm, screening of key genes.}
\usage{
penalty_function_L1(scores, penalty_factor)
}
\arguments{
\item{scores, penalty_factor}{}
}
\value{
Key similar genes
}
\description{
The inner function is used for L1 regularization with cosine similarity algorithm, screening of key genes.
}
\keyword{internal}
