% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{TOPGENE}
\alias{TOPGENE}
\title{TOPGENE Screening}
\usage{
TOPGENE(matrix, threshold = 0.5, genenumber = 200)
}
\arguments{
\item{matrix}{A numeric matrix representing gene expressions.}

\item{threshold}{A numeric value setting the threshold for gene selection.}

\item{genenumber}{An integer indicating the number of top genes to select.}
}
\value{
A matrix containing the selected top genes.
}
\description{
This function uses cosine similarity to screen key genes.
}
\examples{
data <- matrix(rnorm(100), 10, 10)
selected_genes <- TOPGENE(data, threshold = 0.5, genenumber = 5)
}
