% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clGENE.R
\name{select_top_n}
\alias{select_top_n}
\title{This function internally for screening key genes.}
\usage{
select_top_n(scores, n_top)
}
\arguments{
\item{scores, }{n_top}
}
\value{
Key similar genes
}
\description{
This function internally for screening key genes.
}
\keyword{internal}
